chore: splitting files for clarity
