+++ packages/ui/misc/Image/index.tsx
@@ -1,6 +1,13 @@
 import Head from 'next/head';
 import React, { useCallback } from 'react';
-import { HydrogenImage, ImageProps } from './HydrogenImage';
+import { ExternalImage, ExternalImageProps } from './ExternalImage';
+import { iFixitImageLoader, IMG_SRC_SET_SIZES } from './iFixitUtils';
+import { ShopifyImage, ShopifyImageProps } from './ShopifyImage';
+
+export type HtmlImageProps = React.ImgHTMLAttributes<HTMLImageElement>;
+export type ImageProps<GenericLoaderOpts> =
+   | ShopifyImageProps
+   | ExternalImageProps<GenericLoaderOpts>;
 
 export const Image = React.forwardRef(ImageInner);
 
@@ -30,5 +37,51 @@ function ImageInner<GenericLoaderOpts>(
       [preload]
    );
 
-   return <HydrogenImage ref={ref} preloader={preloader} {...otherProps} />;
+   if (!props.data && !props.src) {
+      throw new Error(`<Image/>: requires either a 'data' or 'src' prop.`);
+   }
+
+   if (process.env.NODE_ENV !== 'production' && props.data && props.src) {
+      console.warn(
+         `<Image/>: using both 'data' and 'src' props is not supported; using the 'data' prop by default`
+      );
+   }
+
+   if (props.data) {
+      return (
+         <ShopifyImage
+            ref={ref}
+            preloader={preloader}
+            style={{
+               maxInlineSize: '100%',
+               blockSize: 'auto',
+               objectFit: 'cover',
+            }}
+            {...props}
+         />
+      );
+   } else {
+      const { src, width, height, alt } = props;
+      const loader = iFixitImageLoader({
+         src,
+         width,
+         height,
+         alt,
+         widths: IMG_SRC_SET_SIZES,
+      });
+
+      return (
+         <ExternalImage
+            ref={ref}
+            preloader={preloader}
+            loader={loader}
+            style={{
+               maxInlineSize: '100%',
+               blockSize: 'auto',
+               objectFit: 'cover',
+            }}
+            {...props}
+         />
+      );
+   }
 }
